\name{rdBART-package}
\alias{rdBART-package}
\alias{rdBART}
\docType{package}
\title{
\packageTitle{rdBART}
}
\description{
\packageDescription{rdBART}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rdBART}
\packageIndices{rdBART}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rdBART}

Maintainer: \packageMaintainer{rdBART}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
